dockerfile and run.sh updates to try and make things work on Defang

+++ samples/python/llama/defang-llama/run.sh
@@ -6,6 +6,8 @@ echo $MODEL_FILE
 echo "Model url"
 echo $MODEL_URL
 
-curl -L $MODEL_URL -o "/app/models/${MODEL_FILE}"
+if [ ! -f "/app/models/${MODEL_FILE}" ]; then
+    curl -L $MODEL_URL -o "/app/models/${MODEL_FILE}"
+fi
 
 poetry run flask run
